Update to creation.md
I've replaced the alert TODO with a clear indication that the documentation is simply not ready yet. Defining modules that are accessed via a relative require/import path is not a well covered area and to find some mention of it in the docs but be presented with an alert box isn't great.